Update to the latest Eclipse release 2022-09 (4.25) and to Tycho 3.0.0
Using the Eclipse platform 2022-09 (4.25) and JUnit 5 tests requires
that in the target platform installable units starting with
"org.junit.jupiter." have to be renamed to ones starting with
"junit-jupiter-". This would also apply to existing JUnit Jupiter
references in "Require-Bundle" statements and of features, which is not
the case here, as "Import-Package" statements are used here as
recommended.

This in turn requires Tycho to be updated to 3.0.0 in order to still
recognize JUnit 5 tests by the JUnit Jupiter bundles that have been
renamed in the Eclipse 2022-09 (4.25) release. Otherwise, the build
would fail with the error message "No tests found".

Tycho 3.0.0, in turn again, requires Maven to be run with Java 17 or
higher and the group ID of "tycho-buildtimestamp-jgit" to be changed
from "org.eclipse.tycho.extras" to "org.eclipse.tycho".

See also:
https://github.com/eclipse-tycho/tycho/blob/master/RELEASE_NOTES.md#migration-guide-2x---3x